fix guest stack creation; fix stacks deletion in e2e tests
